componentize talking distribution
